aclocal.m4: Fix BASH_FUNC_SBRK when cross-compiling for FreeBSD/RISC-V
FreeBSD unconditionally provides a prototype for sbrk(2), but on recent
ports (AArch64 and RISC-V) it does not provide the deprecated symbol.
This means that, when cross-compiling, we currently only get to use
AC_CHECK_FUNCS_ONCE, and so believe sbrk(2) to be available. Instead,
use AC_TRY_LINK so we can detect this case when cross-compiling. Also,
only define HAVE_SBRK when our tests pass, since xmalloc.c checks
whether it's defined, not its value. This likely meant the case where
sbrk(2) was available but broken did not in fact do the right thing.