unwind_prot.c: Ensure SAVED_VAR doesn't tag clear pointers
unwind_protect_pointer is a thing that exists and will end up putting
pointers into this structure. I haven't managed to trip up on this yet,
but presumably it's lurking (and jdw57_memcpy on MIPS picks up on a lone
pointer having its tag stripped that goes away with this patch).